Extract http/memory providers to separate packages
