+++ client/client_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/achilleasa/usrv/encoding"
 	"github.com/achilleasa/usrv/server"
 	"github.com/achilleasa/usrv/transport"
-	"github.com/achilleasa/usrv/transport/provider"
+	"github.com/achilleasa/usrv/transport/memory"
 )
 
 func TestClientOptionError(t *testing.T) {
@@ -24,7 +24,7 @@ func TestClientOptionError(t *testing.T) {
 }
 
 func TestClientRequest(t *testing.T) {
-	tr := provider.NewInMemory()
+	tr := memory.NewInMemory()
 	defer tr.Close()
 
 	expGreeting := "hello tester"
@@ -81,7 +81,7 @@ func TestClientRequest(t *testing.T) {
 }
 
 func TestClientRequestWithServerEndpointCtx(t *testing.T) {
-	tr := provider.NewInMemory()
+	tr := memory.NewInMemory()
 	defer tr.Close()
 
 	expSender := "foo service"
@@ -161,7 +161,7 @@ func TestClientMiddlewareChain(t *testing.T) {
 	}()
 	ClearGlobalMiddlewareFactories()
 
-	tr := provider.NewInMemory()
+	tr := memory.NewInMemory()
 	defer tr.Close()
 
 	expReceiver := "service"
@@ -253,7 +253,7 @@ func TestClientMiddlewareChain(t *testing.T) {
 }
 
 func TestClientMiddlewareThatAbortsRequestExecution(t *testing.T) {
-	tr := provider.NewInMemory()
+	tr := memory.NewInMemory()
 	defer tr.Close()
 
 	logChan := make(chan string, 8)
@@ -293,7 +293,7 @@ func TestClientMiddlewareThatAbortsRequestExecution(t *testing.T) {
 }
 
 func TestClientErrors(t *testing.T) {
-	tr := provider.NewInMemory()
+	tr := memory.NewInMemory()
 	defer tr.Close()
 
 	var invocation int32
