+++ transport/http/http.go
@@ -1,4 +1,4 @@
-package provider
+package http
 
 import (
 	"bytes"
