+++ transport/memory/in_memory.go
@@ -1,5 +1,5 @@
 // Package provider contains implementations for the built-in usrv transports.
-package provider
+package memory
 
 import (
 	"fmt"
